+++ src/compiler/parser/mod.rs
@@ -1,0 +1,12 @@
+mod lines_0_1000;
+pub use lines_0_1000::*;
+mod lines_1000_2000;
+pub use lines_1000_2000::*;
+mod lines_2000_3000;
+pub use lines_2000_3000::*;
+mod lines_3000_4000;
+pub use lines_3000_4000::*;
+mod lines_4000_6000;
+pub use lines_4000_6000::*;
+mod lines_6000_8000;
+pub use lines_6000_8000::*;
